requirements.txt
attrs==23.2.0
blinker==1.8.2
bson==0.5.10
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
numpy==1.26.4
openmeteo_requests==1.2.0
openmeteo_sdk==1.11.4
pandas==2.2.2
platformdirs==4.2.2
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-cache==1.2.0
retry-requests==2.0.0
scikit-learn==1.3.0
scipy==1.13.0
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1
Werkzeug==3.0.3
xgboost==2.0.3
